requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
deprecation==2.1.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
gotrue==2.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
postgrest==0.18.0
propcache==0.2.0
pydantic==2.10.0
pydantic_core==2.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==2.0.6
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
yarl==1.17.2
